requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
blinker==1.5
boto3==1.26.50
botocore==1.29.50
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
flask-restx==1.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.1
Mako==1.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
platformdirs==2.6.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.14
virtualenv==20.17.1
Werkzeug==2.2.2
